+++ pheonix-react/package.json
@@ -9,9 +9,11 @@
     "axios": "^1.6.8",
     "bootstrap": "^5.3.3",
     "bootswatch": "^5.3.3",
+    "date-fns": "^3.6.0",
     "firebase": "^10.11.1",
     "lodash": "^4.17.21",
     "react": "^18.2.0",
+    "react-datepicker": "^6.9.0",
     "react-dom": "^18.2.0",
     "react-icons": "^5.1.0",
     "react-router": "^6.23.0",
