+++ pheonix-react/src/App.js
@@ -9,6 +9,7 @@ import Cart from './components/store/Cart';
 import Gift from './components/store/Gift';
 import Purchase from './components/store/Purchase';
 import PurchaseComplete from './components/store/PurchaseComplete';
+import NewMovie from './components/admin/NewMovie';
 
 function App() {
   return (
@@ -19,6 +20,7 @@ function App() {
     <Routes>
               <Route path='/adminMovie' element={<AdminMovie/>}/>
               <Route path='/movieEdit/:movieNo' element={<MovieEdit/>}/>
+              <Route path='/newMovie' element={<NewMovie/>}/>
               <Route path='/login' element={<Login/>}/>
               <Route path='/store' element={<Store />}/>
               <Route path='/cart' element={<Cart />}/>
