+++ pheonix-react/src/components/admin/AdminMovie.css
@@ -36,8 +36,8 @@
 }
 
 .img-preview {
-    width: 100%; /* 너비를 100%로 설정 */
-    height: 200px; /* 높이 설정 */
+    width: 300px; /* 너비를 100%로 설정 */
+    height: 400px; /* 높이 설정 */
     overflow: hidden; /* 넘치는 이미지는 숨김 */
     display: flex; /* 이미지를 가운데 정렬하기 위해 flex 사용 */
     justify-content: center;
@@ -47,4 +47,18 @@
 .img-preview img {
     width: auto; /* 이미지의 원래 비율을 유지 */
     height: 100%; /* 높이는 미리보기 영역에 맞춤 */
+}
+
+.datePicker {
+    display: flex;
+    align-items: center;
+    border-style: none;
+    border-radius: 10px;
+    background-color: rgb(213, 34, 14);
+    box-sizing: border-box;
+    width: 100%;
+    height: 46px;
+    color: white;
+    text-align: center;
+    padding-right: 10px;
 }
